requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.6
djangorestframework==3.15.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
